+++ TestProject/core/src/Objects/GameObject/Entity.java
@@ -165,7 +165,7 @@ public abstract class Entity extends GameObject {
         return this.speed;
     }
 
-    public void setSpeed(float speed) {
+    public void setSpeed(int speed) {
         this.speed = speed;
     }
